Flush output streams before restoring console code pages
When `exit()` is called,
- all functions registered by `atexit()` are called first,
- and then all streams with unwritten buffered data are flushed.

See 'the exit function' in the C standards.
- C89/C90: https://web.archive.org/web/20200909074736if_/https://www.pdf-archive.com/2014/10/02/ansi-iso-9899-1990-1/ansi-iso-9899-1990-1.pdf
- C99: https://open-std.org/JTC1/SC22/WG14/www/docs/n1256.pdf
- C11: https://open-std.org/JTC1/SC22/WG14/www/docs/n1570.pdf
- C17: https://open-std.org/JTC1/SC22/WG14/www/docs/n2310.pdf (diff marks version)

So we must flush output streams before restoring console code pages.
Otherwise, unwritten buffered data will be written as the original
console output code page. In this case, any character which is not in
the original console output code page will be replaced with `?`.
- NOTE: This symptom is specific to console output
        and does not occur when redirecting output to a file.